switch from createReducer to createSlice
