+++ src/Redux/books/bookSlice.js
@@ -1,0 +1,26 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const ADD = 'bookStore/books/ADD';
+const REMOVE = 'bookStore/books/REMOVE';
+
+const addBook = (book) => ({
+  type: ADD,
+  payload: book,
+});
+
+const removeBook = (id) => ({
+  type: REMOVE,
+  id,
+});
+
+const bookSlice = createSlice({
+  name: 'books',
+  initialState: [],
+  reducers: {
+    [ADD]: (state, action) => [...state, action.payload],
+    [REMOVE]: (state, action) => state.filter((book) => book.id !== action.id),
+  },
+});
+export { addBook, removeBook };
+
+export default bookSlice.reducer;
